+++ Source/Shaders/Materials/PolylineArrowMaterial.glsl
@@ -3,12 +3,8 @@
 #endif
 
 uniform vec4 color;
-
-float getPointOnLine(vec2 p0, vec2 p1, float x)
-{
-    float slope = (p0.y - p1.y) / (p0.x - p1.x);
-    return slope * (x - p0.x) + p0.y;
-}
+uniform vec4 outlineColor;
+uniform float outlineWidth;
 
 czm_material czm_getMaterial(czm_materialInput materialInput)
 {
@@ -16,48 +12,53 @@ czm_material czm_getMaterial(czm_materialInput materialInput)
 
     vec2 st = materialInput.st;
 
+    // Fuzz Factor - Controls blurriness of edges
+    const float fuzz = 1.2;
+
+    float headLength = 0.75;   // as a percentage of the line's apparent width (not length!)
+    float centerAmount = 0.15; // half a percentage of the line's apparent width, maximum value is 0.5.
+
 #ifdef GL_OES_standard_derivatives
-    float base = 1.0 - abs(fwidth(st.s)) * 10.0 * czm_pixelRatio;
+    vec2 stPerPixel = fwidth(st);
+
+    // Outline width in st coordinates
+    vec2 outlineSt = outlineWidth * stPerPixel;
 #else
-    float base = 0.975; // 2.5% of the line will be the arrow head
+    float base = 0.025; // 2.5% of the line's length will be the arrow head
+    float guess = 0.1; // 10% of the line's width will be considered a pixel.
+    vec2 stPerPixel = vec2(base * guess / headLength, guess);
+    vec2 outlineSt = vec2(0.0);
 #endif
 
-    vec2 center = vec2(1.0, 0.5);
-    float ptOnUpperLine = getPointOnLine(vec2(base, 1.0), center, st.s);
-    float ptOnLowerLine = getPointOnLine(vec2(base, 0.0), center, st.s);
+    // Find the start and end of the blur between the head and the rest.
+    float baseLow = max(0.5, 1.0 - headLength * stPerPixel.s / stPerPixel.t);
+    float baseHigh = baseLow + fuzz * stPerPixel.s;
+    float headOrNot = smoothstep(baseLow, baseHigh, st.s);
+    float innerHeadOrNot = smoothstep(baseLow + outlineSt.s, baseHigh + outlineSt.s, st.s);
 
-    float halfWidth = 0.15;
-    float s = step(0.5 - halfWidth, st.t);
-    s *= 1.0 - step(0.5 + halfWidth, st.t);
-    s *= 1.0 - step(base, st.s);
+    // Find if we're on the center line.
+    float distanceFromCenter = abs(st.t - 0.5);
+    float centerHigh = centerAmount + fuzz * stPerPixel.t;
+    float centerOrNot = 1.0 - smoothstep(centerAmount, centerHigh, distanceFromCenter);
+    float innerCenterOrNot = 1.0 - smoothstep(centerAmount - outlineSt.t, centerHigh - outlineSt.t, distanceFromCenter);
 
-    float t = step(base, materialInput.st.s);
-    t *= 1.0 - step(ptOnUpperLine, st.t);
-    t *= step(ptOnLowerLine, st.t);
+    // Find if we're inside the sloped edges of the head.
+    float slope =  0.5 - 0.5 * (st.s - baseLow) / (1.0 - baseLow) - distanceFromCenter;
+    float slopeHalfFuzz = 0.5 * fuzz * max(stPerPixel.s, stPerPixel.t);
+    float slopeOrNot = smoothstep(-slopeHalfFuzz, slopeHalfFuzz, slope);
+    float innerSlopeOffset = outlineSt.t + outlineSt.s * 0.5;
+    float innerSlopeOrNot = smoothstep(-slopeHalfFuzz + innerSlopeOffset, slopeHalfFuzz + innerSlopeOffset, slope);
 
-    // Find the distance from the closest separator (region between two colors)
-    float dist;
-    if (st.s < base)
-    {
-        float d1 = abs(st.t - (0.5 - halfWidth));
-        float d2 = abs(st.t - (0.5 + halfWidth));
-        dist = min(d1, d2);
-    }
-    else
-    {
-        float d1 = czm_infinity;
-        if (st.t < 0.5 - halfWidth && st.t > 0.5 + halfWidth)
-        {
-            d1 = abs(st.s - base);
-        }
-        float d2 = abs(st.t - ptOnUpperLine);
-        float d3 = abs(st.t - ptOnLowerLine);
-        dist = min(min(d1, d2), d3);
-    }
+    // Choose centerOrNot vs. slopeOrNot based on headOrNot.
+    float value = mix(centerOrNot, slopeOrNot, headOrNot);
+    float innerValue = mix(innerCenterOrNot, innerSlopeOrNot, innerHeadOrNot);
 
-    vec4 outsideColor = vec4(0.0);
-    vec4 currentColor = mix(outsideColor, color, clamp(s + t, 0.0, 1.0));
-    vec4 outColor = czm_antialias(outsideColor, color, currentColor, dist);
+    // End cap at the back
+    innerValue *= smoothstep(outlineSt.s, outlineSt.s + fuzz * stPerPixel.s, st.s);
+
+    // Color the outline, then the inner line.
+    vec4 outColor = vec4(outlineColor.rgb, outlineColor.a * value);
+    outColor = mix(outColor, color, innerValue);
 
     outColor = czm_gammaCorrect(outColor);
     material.diffuse = outColor.rgb;